+++ src/lib/brain/executive/engines/index.ts
@@ -1,0 +1,17 @@
+export { createTripMonitorEngine } from './tripMonitor'
+export { createLiveConciergeEngine } from './liveConcierge'
+export { createExplainableDecisionEngine } from './explainableDecision'
+export { createTravelMemoryEngine } from './travelMemory'
+export {
+  createMultimodalDocumentEngine,
+  extractFields,
+  detectKind,
+  looksLikeDocument,
+} from './multimodalDocument'
+export type { ExtractedDocumentFields } from './multimodalDocument'
+export { createBudgetIntelligenceV2Engine, estimateBreakdown } from './budgetIntelligenceV2'
+export type { BudgetBreakdown } from './budgetIntelligenceV2'
+export { createItineraryOptimizerEngine } from './itineraryOptimizer'
+export { createRiskEngine } from './riskEngine'
+export { createExecutiveResponseEngine, composeExecutiveReply } from './executiveResponse'
+export { createLearningEngine } from './learningEngine'